Update scr_enemy_ai_b.gml
Changes to disposition system: Upon establishing a connection with a star system, it becomes seamlessly integrated into your sphere of influence. The pivotal determinant here is the 'Visited' flag, serving as the pivotal switch. The status of control is contingent on the faction presiding over the planet. Allowing an incursion onto the planet leads to a reduction in disposition, amounting to 5 in the case of the nurses, and a more significant 10 for the aficionados of all mighty toaster (mechanicus forge worlds). Imperium on the other hand don't give 2 fucks. Planet disposition is independent and will always decrease by 10 if you allow the planet to get overrun.